Refactor facility rendering components for DRY
- Created reusable FacilityAccordion component to handle both library and academic facilities.
- Extracted LibraryRoomsAccordion and AcademicRoomsAccordion for specific room rendering logic.
- Moved RoomBadge to a separate component for better reusability.
- Improved code organization and modularity, reducing duplication across facility types.
- Enhanced maintainability for future expansions and filtering/sorting capabilities.
